+++ src/app.js
@@ -5,7 +5,7 @@ import ReactDOM from 'react-dom';
 import FastClick from 'fastclick';
 import Router from './routes';
 import Location from './core/Location';
-import { addEventListener, removeEventListener } from './utils/DOMUtils';
+import { addEventListener, removeEventListener } from './core/DOMUtils';
 
 let cssContainer = document.getElementById('css');
 const appContainer = document.getElementById('app');